Centralize configuration files
Centralizes all the configuration files of the components used by our
workflows into this repository.

This has the benefits of reducing clutter at the root of the repository
and especially of having everything in one place and allow the
developers to copy/paste the content of this repository when creating
new ones.